pyproject.toml
[project]
name = "test-pg-vector-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"modelscope>=1.30.0",
	"sentence-transformers",
	"xinference>=1.10.0",
	"psycopg2-binary>=2.9",
]
